// src/midtrans-payment/enum/payment-channel.enum.ts
export enum PaymentChannel {
  CREDIT_CARD = 'credit_card',
  GOPAY = 'gopay',
  SHOOPE_PAY = 'shopeepay',
  PERMATA = 'permata_va',
  BCA = 'bca_va',
  BNI = 'bni_va',
  BRI = 'bri_va',
  CIMBNIAGA = 'cimb_va',
  OTHER_VA = 'other_va',
  ECHANNEL = 'echannel',
  ALFAMART = 'alfamart',
}

// available payment channle
// [
//   'gopay',      'qris',
//   'bca_va',     'echannel',
//   'bni_va',     'bri_va',
//   'permata_va', 'cimb_va',
//   'other_va',   'credit_card',
//   'shopeepay',  'qris',
//   'other_qris', 'alfamart',
//   'indomaret',  'akulaku',
//   'kredivo'
// ]